# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi To-do List",
        "description": "RESTful API for a task management application (to-do list).",
        "contact": {
            "email": "luaramineiro@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Demo API To-do List Server"
        }
    ],
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "To-do List"
                ],
                "description": "Returns the list of all tasks.",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Task"
                                        }
                                    ]
                                },
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "current_page": 1,
                                            "data": [
                                                {
                                                    "id": "99f89841-8f88-4ae0-af13-11e934ccf40a",
                                                    "title": "Title example",
                                                    "description": "A description example",
                                                    "attachment": null,
                                                    "is_completed": true,
                                                    "completed_at": "2023-08-25 02:17:34",
                                                    "created_at": "2023-08-25T02:17:33.000000Z",
                                                    "updated_at": "2023-08-25T02:17:33.000000Z",
                                                    "deleted_at": null,
                                                    "user_id": "224ee47b-11cc-4965-b9cd-8464916ce5ad"
                                                }
                                            ],
                                            "links": [
                                                {
                                                    "url": null,
                                                    "label": "&laquo; Previous",
                                                    "active": false
                                                },
                                                {
                                                    "url": "http://localhost:8000/api/tasks?page=1",
                                                    "label": "1",
                                                    "active": true
                                                },
                                                {
                                                    "url": null,
                                                    "label": "Next &raquo;",
                                                    "active": false
                                                }
                                            ],
                                            "next_page_url": null,
                                            "path": "http://localhost:8000/api/tasks",
                                            "per_page": 15,
                                            "prev_page_url": null,
                                            "to": 2,
                                            "total": 2
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "status": "500",
                                            "message": "error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "status": "500",
                                            "message": "error message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "To-do List"
                ],
                "description": "Creates a new task.",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                ]
                            },
                            "examples": {
                                "example": {
                                    "summary": "An example object.",
                                    "value": {
                                        "title": "Title example",
                                        "description": "A description example",
                                        "is_completed": true,
                                        "attachment": [
                                            "file:website.com/pathtofile/intro.pdf",
                                            "file:website.com/pathtofile/photo.png"
                                        ],
                                        "user_id": "224ee47b-11cc-4965-b9cd-8464916ce5ad"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/Task"
                                        }
                                    ]
                                },
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "status": "201",
                                            "message": "Task created",
                                            "data": {
                                                "id": "99f857d3-3119-4be9-b078-c2136ace0314",
                                                "title": "Title example",
                                                "description": "A description example",
                                                "attachment": [
                                                    "file:website.com/pathtofile/intro.pdf",
                                                    "file:website.com/pathtofile/photo.png"
                                                ],
                                                "is_completed": false,
                                                "is_deleted": false,
                                                "deleted_at": null,
                                                "completed_at": null,
                                                "created_at": "2023-08-24T23:17:24.000000Z",
                                                "updated_at": "2023-08-24T23:17:24.000000Z",
                                                "user_id": "224ee47b-11cc-4965-b9cd-8464916ce5ad"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "status": "500",
                                            "message": "error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "status": "500",
                                            "message": "error message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Task": {
                "title": "Task model",
                "description": "Task model",
                "required": [
                    "id",
                    "title"
                ],
                "xml": {
                    "name": "Task"
                }
            }
        }
    },
...
}